+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0 - Initial Release (2025-02-28)
+- ✅ Updated `LionInput` and `LionButton`
+- ✅ Updated UI for `main.dart`
+- 📄 Added Accessbility for Button and Input
+
 ## 0.1.0 - Initial Release (2025-02-28)
 - 🚀 First release of `flutter_lion_components`
 - ✅ Added `LionInput`, `LionButton`, `LionForm`, and `LionDialog` components